Grouped transformations

Grouped transformations

ContinuousTransformations.get_transformation — Function.

get_transformation(d)

Return the transformation from a wrapper object, eg TransformLogLikelihood.

source

ContinuousTransformations.TransformationTuple — Type.

TransformationTuple(transformations::Tuple)
TransformationTuple(transformations...)

A tuple of ContinuousTransformations. Given a vector of matching length, each takes as many reals as needed, and returns the result as a tuple.

source

ContinuousTransformations.TransformLogLikelihood — Type.

TransformLogLikelihood(ℓ, transformations::Union{Tuple, TransformationTuple})
TransformLogLikelihood(ℓ, transformations...)

Return a callable that

transforms its vector argument using a transformation tuple (or a tuple of

transformations, converted as required) to a tuple of values,

calls ℓ with these, which should return a scalar,

returns the result above corrected by the log Jacobians.

Useful when ℓ is a log-likelihood function with a restricted domain, and transformations is used to trasform to this domain from $ℝ^n$.

See also get_transformation.

source