Wrapped transformations

Wrapped transformations

ContinuousTransformations.TransformationWrapper — Type.

abstract TransformationWrapper

Wrap a transformation to achieve some specialized functionality.

Supports length, get_transformation, and other methods depending on the subtype.

source

ContinuousTransformations.TransformLogLikelihood — Type.

TransformLogLikelihood(ℓ, transformation::Union{Tuple, GroupedTransformation})

TransformLogLikelihood(ℓ, transformations...)

Return a callable that

transforms its vector argument using a grouped transformation to a set of values,

calls ℓ with these, which should return a scalar,

returns the result above corrected by the log Jacobians.

Useful when ℓ is a log-likelihood function with a restricted domain, and transformations is used to trasform to this domain from $ℝ^n$.

See also get_transformation, get_distribution, Distributions.logpdf, and logpdf_in_domain.

source

ContinuousTransformations.get_loglikelihood — Function.

get_loglikelihood(t)

Return the log likelihood function.

source

ContinuousTransformations.TransformDistribution — Type.

TransformDistribution(distribution, transformation)

Given a transformation and a distribution, create a transformed distribution object that has the distribution of transformation(x) with x ∼ distribution.

It supports logpdf, rand, length. The transformation object is callable and works the same way as t.

See also logpdf_in_domain for calculating the log pdf from the untransformed values.

source

ContinuousTransformations.get_distribution — Function.

get_distribution(t)

Return the wrapped distribution.

source

ContinuousTransformations.logpdf_in_domain — Function.

logpdf_in_domain(t, x)

For a transformed distribution which maps x using a transformation, return the log pdf for a given x. The log pdf is adjusted with the log determinant of the Jacobian, ie the following holds:

logpdf(t, t(x)) == logpdf_in_domain(t, x)

See Distributions.logpdf.

source